+++ plugins/rest_web/RestWebPlugin.cc
@@ -20,6 +20,8 @@
   #pragma comment(lib, "Rpcrt4.lib")
 #else /* UNIX */
 
+#include <gazebo/gazebo.hh>
+
 #ifdef HAVE_UUID
   #include <uuid/uuid.h>
 #endif
@@ -67,7 +69,11 @@ RestWebPlugin::RestWebPlugin()
 #endif
 
 #endif
-
+  if (this->session.empty())
+  {
+    // alternative to uuid
+    this->session = common::Time::GetWallTimeAsISOString();
+  }
   gzmsg << "REST web Session : " << this->session << endl;
 }